+++ openbb_platform/extensions/stocks/integration/test_stocks_api.py
@@ -681,6 +681,24 @@ def test_stocks_options_chains(params, headers):
 @pytest.mark.parametrize(
     "params",
     [
+        ({"symbol": None, "source": "delayed", "provider": "intrinio"}),
+        ({"symbol": "PLTR", "source": "delayed", "provider": "intrinio"}),
+    ],
+)
+@pytest.mark.integration
+def test_stocks_options_unusual(params, headers):
+    params = {p: v for p, v in params.items() if v}
+
+    query_str = get_querystring(params, [])
+    url = f"http://0.0.0.0:8000/api/v1/stocks/options/unusual?{query_str}"
+    result = requests.get(url, headers=headers, timeout=10)
+    assert isinstance(result, requests.Response)
+    assert result.status_code == 200
+
+
+@pytest.mark.parametrize(
+    "params",
+    [
         (
             {
                 "symbol": "AAPL",