+++ openbb_platform/platform/provider/openbb_provider/standard_models/etf_countries.py
@@ -26,7 +26,6 @@ class EtfCountriesQueryParams(QueryParams):
 class EtfCountriesData(Data):
     """ETF Countries Data."""
 
-    country: str = Field(description="The country of the exposure.")
-    weight: float = Field(
-        description="Exposure of the ETF to the country in normalized percentage points."
+    country: str = Field(
+        description="The country of the exposure.  Corresponding values are normalized percentage points."
     )